refactor: extract shared logic for link checking
